+++ plato_calendar/lib/Page/widget/appointmentEditor.dart
@@ -6,7 +6,6 @@ import 'package:syncfusion_flutter_calendar/calendar.dart';
 
 import '../../Data/subjectCode.dart';
 import '../../Data/userData.dart';
-import '../../Data/database/database.dart';
 import '../../Data/ics.dart';
